Updated SearchTab UI
