+++ lib/screens/search_tab/search_tab.dart
@@ -1,183 +1,97 @@
-import 'package:anonaddy/models/alias/alias.dart';
-import 'package:anonaddy/screens/alias_tab/alias_screen.dart';
-import 'package:anonaddy/screens/alias_tab/components/alias_shimmer_loading.dart';
-import 'package:anonaddy/screens/search_tab/components/search_history.dart';
-import 'package:anonaddy/screens/search_tab/components/search_list_header.dart';
-import 'package:anonaddy/screens/search_tab/components/search_text_field.dart';
+import 'package:anonaddy/notifiers/search/search_history/search_history_notifier.dart';
+import 'package:anonaddy/notifiers/search/search_history/search_history_state.dart';
 import 'package:anonaddy/shared_components/constants/constants_exports.dart';
 import 'package:anonaddy/shared_components/list_tiles/alias_list_tile.dart';
-import 'package:anonaddy/shared_components/lottie_widget.dart';
 import 'package:anonaddy/shared_components/platform_aware_widgets/platform_scroll_bar.dart';
-import 'package:anonaddy/notifiers/search/search_history/search_history_notifier.dart';
-import 'package:anonaddy/notifiers/search/search_result/search_result_notifier.dart';
-import 'package:anonaddy/notifiers/search/search_result/search_result_state.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-class SearchTab extends StatelessWidget {
+class SearchTab extends ConsumerWidget {
   const SearchTab({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context) {
-    final size = MediaQuery.of(context).size;
-
-    return GestureDetector(
-      onTap: () => FocusScope.of(context).requestFocus(FocusNode()),
-      child: Scaffold(
-        body: NestedScrollView(
-          physics: const ClampingScrollPhysics(),
-          headerSliverBuilder: (context, _) {
-            return [
-              SliverAppBar(
-                expandedHeight: size.height * 0.21,
-                elevation: 0,
-                flexibleSpace: FlexibleSpaceBar(
-                  collapseMode: CollapseMode.pin,
-                  background: SearchTabHeader(),
+  Widget build(BuildContext context, WidgetRef ref) {
+    return Scaffold(
+      body: Column(
+        children: [
+          Container(
+            color: AppColors.primaryColor,
+            padding: const EdgeInsets.symmetric(horizontal: 10),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Text(
+                  AppStrings.searchHistory,
+                  style: Theme.of(context)
+                      .textTheme
+                      .headline6
+                      ?.copyWith(color: Colors.white),
                 ),
-              ),
-            ];
-          },
-          body: Consumer(
-            builder: (context, ref, _) {
-              final searchState = ref.watch(searchResultStateNotifier);
+                TextButton(
+                  child: const Text(
+                    'Clear',
+                    style: TextStyle(color: Colors.red),
+                  ),
+                  onPressed: () {
+                    ref
+                        .read(searchHistoryStateNotifier.notifier)
+                        .clearSearchHistory();
+                  },
+                ),
+              ],
+            ),
+          ),
+          Consumer(
+            builder: (context, ref, child) {
+              final searchState = ref.watch(searchHistoryStateNotifier);
 
               switch (searchState.status) {
-
-                /// When search is NOT triggered, basically the default state for [SearchTab]
-                /// In this case, show [SearchHistory].
-                case SearchResultStatus.initial:
-                  return Column(
-                    children: [
-                      SearchListHeader(
-                        title: AppStrings.searchHistory,
-                        buttonLabel: AppStrings.clearSearchHistoryButtonText,
-                        buttonTextColor: Colors.red,
-                        onPress: () {
-                          ref
-                              .read(searchHistoryStateNotifier.notifier)
-                              .clearSearchHistory();
-                        },
-                      ),
-                      const SearchHistory(),
-                    ],
+                case SearchHistoryStatus.loading:
+                  return const Padding(
+                    padding: EdgeInsets.all(20),
+                    child: Center(child: CircularProgressIndicator()),
                   );
 
-                /// Displays limited search results based on searching through
-                /// locally available aliases, typically page 1 of user's aliases.
-                case SearchResultStatus.limited:
-                  final aliases = searchState.aliases!;
-                  return buildResult(context, ref, aliases, true);
+                case SearchHistoryStatus.loaded:
+                  final aliases = searchState.aliases;
 
-                /// When search is loading e.g. fetching matching aliases
-                case SearchResultStatus.loading:
-                  return Column(
-                    children: [
-                      SearchListHeader(
-                        title: AppStrings.searching,
-                        buttonLabel: AppStrings.cancelSearchingButtonText,
-                        buttonTextColor: Colors.red,
-                        onPress: () {
-                          /// Close current on-going search
-                          ref
-                              .read(searchResultStateNotifier.notifier)
-                              .closeSearch();
-                        },
+                  if (aliases.isEmpty) {
+                    return Padding(
+                      padding: const EdgeInsets.all(20),
+                      child:
+                          Row(children: const [Text('Nothing to see here.')]),
+                    );
+                  } else {
+                    return Expanded(
+                      child: PlatformScrollbar(
+                        child: ListView.builder(
+                          shrinkWrap: true,
+                          itemCount: aliases.length,
+                          // physics: const NeverScrollableScrollPhysics(),
+                          itemBuilder: (context, index) {
+                            return AliasListTile(alias: aliases[index]);
+                          },
+                        ),
                       ),
-                      const Expanded(child: AliasShimmerLoading()),
-                    ],
-                  );
+                    );
+                  }
 
-                /// When search has finished loading and returns matching aliases
-                case SearchResultStatus.loaded:
-                  final aliases = searchState.aliases ?? [];
-                  return buildResult(context, ref, aliases, false);
-
-                /// When searching fails and returns an error
-                case SearchResultStatus.failed:
-                  final errorMessage = searchState.errorMessage;
-                  return LottieWidget(
-                    label: errorMessage,
-                    lottie: LottieImages.emptyResult,
-                    lottieHeight: 150,
+                case SearchHistoryStatus.failed:
+                  final error = searchState.errorMessage;
+                  return Padding(
+                    padding: const EdgeInsets.all(20),
+                    child: Row(
+                      children: [
+                        Text(error ??
+                            'Something went wrong: ${error.toString()}'),
+                      ],
+                    ),
                   );
               }
             },
           ),
-        ),
+        ],
       ),
-    );
-  }
-
-  Widget buildResult(BuildContext context, WidgetRef ref,
-      List<Alias> resultsList, bool isLimited) {
-    final size = MediaQuery.of(context).size;
-    return Column(
-      children: [
-        SearchListHeader(
-          title: isLimited
-              ? AppStrings.limitedSearchResult
-              : AppStrings.searchResult,
-          buttonLabel: AppStrings.closeSearchButtonText,
-          onPress: () {
-            FocusScope.of(context).requestFocus(FocusNode());
-            ref.read(searchResultStateNotifier.notifier).closeSearch();
-          },
-        ),
-
-        /// Display a banner indicating that current results are from a limited list
-        if (isLimited)
-          Container(
-            padding: const EdgeInsets.symmetric(vertical: 10),
-            child: const Text('Tap Search for a full search'),
-          ),
-
-        /// Display nothing if result from a limited list is empty
-        if (resultsList.isEmpty && isLimited)
-          Container()
-        else
-
-        /// Show empty lottie if result from full search is empty
-        if (resultsList.isEmpty)
-          LottieWidget(
-            lottie: LottieImages.emptyResult,
-            lottieHeight: size.height * 0.12,
-          )
-        else
-          Expanded(
-            child: PlatformScrollbar(
-              child: ListView.builder(
-                shrinkWrap: true,
-                itemCount: resultsList.length,
-                physics: const NeverScrollableScrollPhysics(),
-                itemBuilder: (context, index) {
-                  final alias = resultsList[index];
-                  return InkWell(
-                    child: IgnorePointer(
-                      child: AliasListTile(alias: alias),
-                    ),
-                    onTap: () {
-                      /// Dismisses keyboard
-                      FocusScope.of(context).requestFocus(FocusNode());
-
-                      /// Add selected Alias to Search History
-                      ref
-                          .read(searchHistoryStateNotifier.notifier)
-                          .addAliasToSearchHistory(alias);
-
-                      /// Navigate to Alias Screen
-                      Navigator.pushNamed(
-                        context,
-                        AliasScreen.routeName,
-                        arguments: alias,
-                      );
-                    },
-                  );
-                },
-              ),
-            ),
-          ),
-      ],
     );
   }
 }
